requirements.txt
openai-whisper
anthropic
openai
pyannote.audio
torch
